move `functions` to `src/utils`
